handle scalarized constraints in ResultsAnalysis (#4193)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4193

`ResultsAnalysis` will err when trying to extract metric names from `ScalarizedOutcomeConstraint`. This change fixes it.

Meta: Currently we treat the mean/sem of the scalarized outcome constraint as nan (hence the feasibility of it) and avoid the error. Created T235432214 explicitly calculate the scalarized constraint values when calculating feasibility.

Reviewed By: Balandat

Differential Revision: D80653847

fbshipit-source-id: 596c9489cb1ebab8b10094ca29d064a33a48a1ab