Handle multiple image paths in same md doc (#4198)
Summary:
X-link: https://github.com/pytorch/botorch/pull/2992

The ipynb to mdx conversion script has a step to copy images to the mdx destination and update the relative image paths. The issue was that modifying the md buffer causes the static regex match indices to become out of date.

The fix here here is to process the matches in reverse order, that way modifications at the end of the buffer don't affect the indices for the next replacements (which are earlier in the buffer). This is the same solution used in the `handle_image_attachments()` function which experienced a similar problem

eonofrey found this issue while adding images to a tutorial in f2186fd9fbbfc403c443e80d7cbc5011f9bb376b

| Before    | After |
| -------- | ------- |
|        <img width="1290" height="722" alt="image" src="https://github.com/user-attachments/assets/8c68e78c-c78e-4c61-87e7-3610a2229372" />         |     <img width="1296" height="1308" alt="image" src="https://github.com/user-attachments/assets/f5abea10-af97-41f7-9a9c-8c21c28d7a25" />       |


Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Rollback Plan:

Reviewed By: eonofrey

Differential Revision: D80961417

Pulled By: CristianLara