Exclude out-of-design arms at dedup (#4216)
Summary:

f784259720 errored out when `RandomAdapter` is trying to transform arms to dedup, which includes a custom arm that is out-of-design.

As a fix, we filter out-of-design arms at dedup to avoid failure at `Transform`. Since we will not actually generate any arms that are out-of-design, this filtering should be safe / is only used for custom arms.

Reviewed By: saitcakmak

Differential Revision: D81171470