Add trial-based cross validation (#4214)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4214

Suppose I fit a model and use it to generate a trial. Afterwards I may be interested to know if the model predictions that formed the basis of that trial actually match the observed outcomes.

The cross validation flow is a natural place for this type of predict vs. observed comparison. This diff refactors cross_validate to enable plugging in different train/test generators besides the current arm-based k-fold. It also implements another fold generator based on trial index, which can be used for the purpose described above.

Reviewed By: saitcakmak

Differential Revision: D81061406