Remove unused parameters input from TorchAdapter._fit, _cross_validate & _get_fit_args
Summary:
Allowing a different set of parameters than the parameters available on the search space can lead to inconsistencies between model fitting and subsequent use of the fitted model. Since this argument is not used, let's just remove it.

After removing it from `_fit` and `_cross_validate`, it becomes `None` everywhere for `_get_fit_args` as well, so it is removed from there too.

Reviewed By: dme65

Differential Revision: D81511134