Remove unused parameters input from TorchAdapter._fit, _cross_validate & _get_fit_args (#4234)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4234

Allowing a different set of parameters than the parameters available on the search space can lead to inconsistencies between model fitting and subsequent use of the fitted model. Since this argument is not used, let's just remove it.

After removing it from `_fit` and `_cross_validate`, it becomes `None` everywhere for `_get_fit_args` as well, so it is removed from there too.

Reviewed By: dme65

Differential Revision: D81511134

fbshipit-source-id: 2cf5ce0f3045b7ca5227927fa20dd7a0dfa8e829