Weighted & batched EnsemblePosterior Ax support (#4201)
Summary:
Adds more predict support for `EnsemblePosterior`, and makes it more coherent with the predict support that is already in place for `GaussianMixturePosterior` by using the newly added (previous diff) `mixture` attributes for `EnsemblePosterior`.

In the case where an `EnsemblePosterior` has batch dimensions in addition to the ensemble dimension, we compute the mean and variance over both batch and ensemble dimensions, instead of just the latter. Moreover, this allows us to properly account for non-uniform weights across batch and ensemble dimensions, which is needed for sampling-based benchmarking in ensembled Fully Bayesian models.

Removed tests that involved previous mixture moment calculations, as these are now in BoTorch.


Reviewed By: saitcakmak

Differential Revision: D80972578