Back out "allow different inferred noise levels for each task in MultitaskGP" (#4253)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4253

Only backing out portions of this diff which touch Ax.

We intend to release Ax 1.1.1 soon, but are currently unable to without releasing new gpytorch and botorch versions due to the inclusion of HadamardGaussianLikelihood.
In the interest of time, we can remove references to HadamardGaussianLikelihood from Ax, release Ax 1.1.1 without moving the botorch pin from 0.15.1, and quickly re-add this test with the intention of releasing a new botorch and gpytorch version before Ax 1.1.2

Reviewed By: Balandat

Differential Revision: D81695384

fbshipit-source-id: 4ab2d76b83f2fd679bda80782c08404d4cb114ac