Remove MapKeyInfo, making all map keys "step" (#4255)
Summary:

With this change, a map key can no longer be specified on a MapMetric or MapData; the key will be "step." D81254898 will clean up various references to map keys.

* Remove MapKeyInfo (no longer an attribute on MapData or MapMetric)
* All map keys become "step" by default, so map keys were renamed in various places
* As needed, I either updated various references to `MapKeyInfo`s to point to the default map key or changed them to None (in cases in which that will preserve the correct behavior). D81254898 will remove various references to map_key.
* Removed map_key arguments from early stopping functions, since this was easier that doing some half-solution where we pass the default map key for now.

Note: `MapKeyInfo` does not appear to have ever been serialized and stored. It is not used with serialization for `MapMetric` because it is not an init arg, and serialization for `MapData` deals with `MapKeyInfo` only in a serialized format. I added storage backwards compatibility for `MapData`; it should work automatically with `MapMetric`, and I don't think serialization backwards compatibility is needed for `MapKeyInfo` itself. However, this would be a regression if there are custom `MapMetric` subclasses that I haven't found that use `MapKeyInfo` as an init arg.

Reviewed By: Balandat

Differential Revision: D81363344