Enabled posterior samples as SurrogateBenchmarks (#4163)
Summary:

This diff enables posterior samples as SurrogateBenchmarks, by substituting the model for a `DeterministicModel`, which can be either a PosteriorMeanModel or a `GenericDeterministicModel`. `DeterministicModel`s make use of the `EnsemblePosterior`, so predict support for these is added as well. 

This enables the flexibility of choosing to evaluate the posterior mean or a sample path (or any other DeterministicModel) by passing in said model as an optional argument to the `SurrogateTestFunction`.

TL;DR:
- substitutes the model in surrogate benchmarks for a DeterministicModel - default `PosteriorMeanModel`
- `EnsemblePosterior` support in `BoTorchGenerator.predict`
- Allowing the choice of either a PosteriorMean or a MatheronPath (or something else if we come up with it down the road)

Another potential option (I think) would be to use evaluate_acquisition_function, and use either PosteriorMean or PathwiseThompsonSampling.

Reviewed By: saitcakmak

Differential Revision: D80347570