Add Experiment Name to Logs of Stale Failed Trials for Better Debugging (#4303)
Summary:

This diff adds the experiment name to the logs of stale failed trials for better debugging. The code change in the file `fbcode/ax/service/orchestrator.py` adds the experiment name to the log message. This will help in identifying the experiment for which the trials are failing [here](https://fburl.com/rs56b5yu).

Reviewed By: lena-kashtelyan

Differential Revision: D82578274