Add Experiment Name to Logs of Stale Failed Trials for Better Debugging (#4303)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4303

This diff adds the experiment name to the logs of stale failed trials for better debugging. The code change in the file `fbcode/ax/service/orchestrator.py` adds the experiment name to the log message. This will help in identifying the experiment for which the trials are failing [here](https://fburl.com/rs56b5yu).

Reviewed By: lena-kashtelyan

Differential Revision: D82578274

fbshipit-source-id: 9f16dd46b3bebdd2818454f31319f470918874a2