Remove no/use_winsorization option
Summary: With the default set of generation strategies we dispatch to, there is very little benefit to turning off winsorization. These models are not robust to outliers and will regress if winsorization is not used. Removing `no_winsorization` option will simplify the addition of `Winsorize` to `Y_trans` and use of winsorization by default from there on.

Reviewed By: dme65

Differential Revision: D82322529