automatically fall back to the hit and run sampler in RandomGenerator when there are continuous parameters (#4299)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4299

see title.

There is context in this comment on why we didn't do this previously: https://www.internalfb.com/diff/D31431838?dst_version_fbid=611591606519560&transaction_fbid=304592317864246

The TL;DR is that we needed to distinguish (a) when the search space was actually exhausted (in the case of all discrete parameters) from (b) when the search space is not exhausted, but rather the rejection sampling failed due to the constraints being hard to satisfy.

We are always in setting (b) when there is at least one continuous parameter, so we should fallback to the hit and run sampler automatically.

This also re-uses de-duplication logic from rejection sample when using the hit and run sampler.

This came up recently in experiment that andycylmeta was running

Reviewed By: Balandat

Differential Revision: D82250529

fbshipit-source-id: a0b8cd5c1690205474a55134539014c6962a1e62