Remove "fidelities" column from data (#4314)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4314

**Context**:

Fidelities in Ax are handled as parameters, not as part of Data. D16586847 added fidelities to Data, but most of the functionality for working with fidelities has since been deleted. At this point, "fidelities" is a column that is special only in that it is expected to be json-encoded and is json-decoded in `_observations_from_dataframe` (this isn't really documented in Data).

The "fidelities" column in Data does not seem to be used.

**This PR**:
* Removes the "fidelities" column from `Data`

Reviewed By: saitcakmak

Differential Revision: D82750373

fbshipit-source-id: 2dd154841fa7cafff03a3bfe501ab59ffa971a91