Introduce optional `signature` and `label` attributes to Metric, make `name` argument optional (#3864)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3864

As titled - throughout our code and data models, we utilize metric names for storage and referencing purposes. However, dealing with metric names can be challenging, particularly when dealing with lengthy names that are difficult to repeatedly reference.

This change is to introduce the canonical name (signature) and display name (label) for the `Metric` class.

The end goal is to replace the `name` attribute with `signature`. In the envisioned end state, `Metric.signature` will be used under the hood, just as `Metric.name` is currently being used. `Metric.label` will be used for user facing experiences, such as figures, printed outputs, etc.

This change introduces two new optional arguments, `signature` and `label`, to the Metric class. Notably, if `label` is not provided, it will default to the value of `signature`.

We retain the `name` argument, but raise a DeprecationWarning when used. If both `name` and `signature` are provided, we raise an error if the values differ.

This change represents an intermediate step, as we plan to eventually deprecate and remove the existing `name` argument in favor of these new parameters.

Differential Revision: D75476155