Remove support for passing `sample_sizes` (#4315)
Summary:

**Context**:

When `sample_sizes` are passed in Ax, e.g. via `AxClient.complete_trial`, they ultimately become the "n" column in Data. Sample sizes do not have any real use within OSS Ax and are never passed this way (although "n" can be passed directly when creating `Data`).

I think `sample_size` has the potential to be misleading since a user might think Ax will do something with it or that they have to pass it.

**This PR**:

Removes `sample_sizes` from various calls that update trial data. There is no change to the API because this is already not supported in the API.

Reviewed By: saitcakmak

Differential Revision: D82770222