Disable the root check after adding metadata to the search (#4312)
Summary:

`MapKeyToFloat` was broken for hierarchical search spaces when the experiment contains `MapData`. This transform adds new parameters to the search space, and those added parameters have no parents. This would fail the root check in `HierarchicalSearchSpace`, which requires the root to be the only independent parameter.

Note that the same issue also applies to other transforms involving metadata, e.g., `MetadataToFloat` and `MetadataToTask`.

To resolve this issue, this diff modifies `MetadataToParameterMixin` to disable the root check once the parameters are added.

Reviewed By: saitcakmak

Differential Revision: D82679645