Introduce `signature` attribute and optional `signature_override` init argument (#3864)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3864

As titled - throughout our code and data models, we utilize metric names for storage and referencing purposes. However, dealing with metric names can be challenging, particularly when dealing with lengthy names that are difficult to repeatedly reference and to display.

This change is to introduce a metric signature (aka canonical name).

Metric signatures are defaulted to name, and can be overridden via the `signature_override` arg.

The goal is to replace the `name` attribute with `signature`. In the envisioned end state, `Metric.signature` will be used under the hood, just as `Metric.name` is currently being used. `Metric.name` will be used for user facing experiences, such as figures, printed outputs, etc.

Reviewed By: lena-kashtelyan

Differential Revision: D75476155

fbshipit-source-id: 0a9776b17d9308d9a04ab7484d1f8c28aa6800e2