Update test observation transfroms in cross_validate when untransform=False
Summary:
The current implementation transforms `ExperimentData` and extracts the observations from there using the `test_selector`, if appliable. The catch is that the `test_selector` may not be compatible with observations from the transformed space. This may lead to changing the set of observations that are used when constructing the `CVResults`.

This diff replaces the transform on `ExperimentData` objects with transforms on already filtered (by `test_selector)` `cv_test_observations`. This will make sure that the observations that we use for making predictions are consistent with the observations they are being compared to.

Reviewed By: Balandat

Differential Revision: D82984880