Update scoring for constrained problems (#4331)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4331

This diff updates the scoring for constrained problems in the Ax benchmarking based on what was used in the Ax paper. That is, infeasible observations (and potentially the baseline value) are assigned the worst known feasible value when computing the score for constrained problems. This has the desirable property that any feasible trial has better performance than an infeasible trial. This diff also adds a new `is_feasible_trace` to make it easier to understand which trials were feasible.

Reviewed By: esantorella

Differential Revision: D82383753

fbshipit-source-id: 8af09887fcebee99c68c7f0d4f2df389e0862c84