Remove observations from transform constructors in Winsorize tests
Summary:
This is done in a separate diff since the winsorize tests required a lot of updates.

**Context:** With the migration to `ExperimentData` usage in `Transform`, `Adapter` always uses `ExperimentData` when constructing the `Transform` objects. We do not need to support `observations` argument in the `Transform` constructors anymore. Removing it now will simplify the planned work around improved transform configs, so I decided to clean it up.

To ensure actual functionality is not affected, we first update all tests to use `experiment_data` rather than `observations` when initializing the transforms. All usage of `observations` in the constructors is replaced by dummy `Experiment`s and `ExperimentData` extracted from these. The data is kept as similar to original as possible. Some duplicate test have been removed.

Differential Revision: D83094670