Reap non-early-stopping LCBench transfer learning benchmarks
Summary: These are not used, and there is a different set of LCBench surrogates used for early stopping.  If these aren't deleted, the storage needs to be reworked to use Ax serialization instead of torch.load/torch.save, or else changes to the attributes of the serialized classes will prevent them from being loaded properly.

Reviewed By: saitcakmak

Differential Revision: D83098718