Remove support for observations in transform constructors (#4346)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4346

With the migration to `ExperimentData` usage in `Transform`, `Adapter` always uses `ExperimentData` when constructing the `Transform` objects. We do not need to support `observations` argument in the `Transform` constructors anymore. Removing it now will simplify the planned work around improved transform configs, so I decided to clean it up.

Differential Revision: D83095274