Add base Transform to the storage registry
Summary: In the decoder, we replace the deprecated transforms with the base `Transform`. If the replaced `Transform` happens to be stored later (e.g., you loaded the experiment with Client and did some operations that updated in the DB), next time we try to load the GS, we will attempt to decode `Transform`. This errors out since base `Transform` is currently not in the registry.

Differential Revision: D83262852