more fault tolerant mast benchmarking
Summary:
The MAST-calling jobs now have a retry count s.t. they are restarted if they fail e.g. due to manifold errors. Additionally, I have added a timeout of 12 hours s.t. a job that gets stuck fails after that time.

This of course has the downside, that if a job is meant to take this long it will be failed, I'd say we should not really have any jobs like this, though, if we properly parallelize, no? Our nightlies for example take 5 hours at most (which is already a little unnecessarily long, given that we can parallelize things basically ad infinitum, no?)

Differential Revision: D81128135