store and use target point from OptimizationConfig (#4392)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4392

This adds a `target_arm` to optimization config which will be used as the target point for pruning irrelevant parameters. If no `target_arm` is specified on the `OptimizationConfig`, the SQ will be used if it is in the search space.

This throws the target_arm in experiment._properties for SQA storage.

Differential Revision: D83677314