prune irrelevant parameters after generating candidates (#4366)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4366

See title. The method involves first optimizing the AF without any notion of irrelevance. Then, the irrelevant parameters are pruned via a sequential greedy algorithm that sets each dimension to the target point value, and recomputes the AF value. The dimension that corresponds to the smallest reduction in Ax value (compared the AF value of the dense original candidate) is then set to the target value and the algorithm proceeds until removing any dimension causes the reduction of the AF value of the proposed point to be less than some tolerance, relative to the original point.

For q>1 cases, this is repeated for each point in the batch, after adding the previously pruned points as pending points. The stopping rule is applied to the incremental AF value

This first diff adds support for pruning irrelevant parameters where the target_point is passed via `acquisition_options`. A subsequent diff will implement support for storing and passing a target point through Ax transforms, down to the modeling layer.

Reviewed By: saitcakmak

Differential Revision: D83097348

fbshipit-source-id: 39f192a35d7e1b7ef7209d6e7fc03d9b9734f4b2