Don't cast numpy types to Python types in `formatting_utils.py` (#4389)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4389

Currently, in `raw_data_to_evaluation`, Numpy types error when data is put in with a metric name, and they are cast to Python types when data is put in without a metric name. Python ints are cast to floats (except with map data). This is only for the mean; the SEM is ignored.

This is unnecessary because the data will just be used to produce a DataFrame within `Data` and then will be forcibly cast to a float. Pandas handles this just fine (and in a vectorized way), so there is no need for Ax to do this.

Reviewed By: Balandat

Differential Revision: D83847415

fbshipit-source-id: 7e70a107df9979c3f1456a0ba82dcff927814d4e