Update _input_transform_argparse_normalize to support ContextualDataset
Summary: Contextual datasets specify a parameter decomposition, which is used to reshape the input tensors within the underlying model before applying the input transforms. To ensure that the default `Normalize` transform is compatible with such models, we need to construct it using inputs from the transformed space. This diff extracts `d, bounds & indices` that correspond to parameters in a given context before constructing the input transform

Reviewed By: bletham

Differential Revision: D84076532