store and use target point from OptimizationConfig (#4392)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4392

This adds a `target_arm` to optimization config which will be used as the target point for pruning irrelevant parameters. If no `target_arm` is specified on the `OptimizationConfig`, the SQ will be used if it is in the search space.

This puts the `target_arm` in `experiment._properties` for SQA storage. Currently, the OptConfig is cloned without the `target_arm` before storing the copy of the OptConfig on the GeneratorRun.

Reviewed By: lena-kashtelyan

Differential Revision: D83677314

fbshipit-source-id: f14a15956c84c7751a1c6c2c6d6e54521132c67b