Restrict BoTorch version requirement (#4412)
Summary:
We've noticed that a version requirement like `botorch[pymoo]>=0.15.1` might cause issues if we release BoTorch versions before new Ax versions. This diff updates the requirement to allow the development versions but not the full release of the next BoTorch version.

The catch is that once the BoTorch version is released, the Ax GHA workflows will not use it until the pin is updated (which we tend to do right after the release). I think this is a fair compromise.

Pull Request resolved: https://github.com/facebook/Ax/pull/4412

Reviewed By: mpolson64

Differential Revision: D84191006

Pulled By: saitcakmak

fbshipit-source-id: 9eda4dfbfff53d3018c535a3f693d8c591fe5378