(1/3) Move relativize function out of statstools to avoid circular dependency in imports (#4414)
Summary:

This diff is part of a series of changes to resolve a circular dependency between `ax.core` and `statstools`. Currently, the relativize function (in `statstools`) cannot be imported from within the `ax.core` package because `statstools` itself depends on `ax.core`. In this specific diff, the `relativize` function is moved out of the `statstools` module. 
To address this:
- The `relativize` function has been moved from `statstools` to a new file `math_utils`.
- Previously, there were no dedicated unit tests for `relativize`, though its basic functionality was covered by [RelativizeDataTest](https://www.internalfb.com/code/fbsource/[6fbf07b1c6a1826d6e3dbe7c8e3df9bd47b60043]/fbcode/ax/utils/stats/tests/test_statstools.py?lines=81) in test_statstools.py. This diff adds new tests to cover previously untested parts of `relativize`.
- Imports in `ax.utils.stats.statstools` have been updated to reference the new location of `relativize`.

Differential Revision: D84262111