(2/3) Move unrelativize function out of statstools to avoid circular dependency in imports (#4415)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4415

This diff is part of a series of changes to resolve a circular dependency between `ax.core` and `statstools`. Currently, the relativize function (in `statstools`) cannot be imported from within the `ax.core` package because `statstools` itself depends on `ax.core`. In this specific diff, the `unrelativize` function is moved out of the `statstools` module.
To address this:
- The `unrelativize` function has been moved from `statstools` to `math_utils`.
- Move `UnrelativizeTest` from `test_statstools` to `test_math_utils`
- Imports in `ax.utils.stats.statstools` have been updated to reference the new location of `unrelativize`.

Reviewed By: esantorella

Differential Revision: D84262809

fbshipit-source-id: 021c21585eb4cb02555f53e61b4e9eaa8c2143ac