Add generic `botorch_synthetic` problem registry entry (#4433)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4433

Currenlty, any test function we'd want to run benchmarks against would need to be added explicitly to the registry. However, it would be convenient to be able to do this simply by name without having to register each individually, especially since all the botorch test functions are importable anyway.

This diff adds the capability for this by enabling the `test_problem_class` argument in `create_problem_from_botorch` to be a string (in additon to a class). If the test function class name is provided as a string, the class will be loaded dynamically via `getattr` from the module `botorch.test_functions.synthetic`.

Reviewed By: esantorella

Differential Revision: D84467574

fbshipit-source-id: fa5639309358e0c72d2ae4d7bebdbb83c1f5a3d1