Add Analysis.vaidate_applicable_state (#4428)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4428

Add a new method to Analysis which can be used to determine whether the given Experiment, GenerationStrategy, and/or Adapter are in an appropriate state to compute the given analysis. This allows us to centralize a number of checks which currently exist in various `compute` methods and provides a common and obvious place to add future validation.

We currently have a problem with Ax where we attempt to compute analyses while the Experiment, GenerationStrategy, and/or Adapter are not in an appropriate state to produce the analyses, and this infra will help encourage better practices.

Differential Revision: D82996448