Create standalone p_feasible plot (#4443)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4443

We removed the ability to plot p_feasible as a pseudo metric on ArmEffectsPlot. Adding a new PFeasiblePlot which subsumes that functionality using shared utilities.

Performs dedicated validation, and returns a plot with one point per arm and raw or modeld probability of feasibility on the y axis.

Also updated `__repr__` on ScalarizedOutcomeConstraint to be more in line with other representations in the same class.

Reviewed By: sdaulton

Differential Revision: D84827303