Use named arguments when instantiating objects / calling functions
Summary: Asked devmate to find usage of positional arguments and replace them with named arguments. It finds a few classes that are commonly used with positional args & updates those. I ask for more and it repeats. After a bunch of iterations, we end up with this. Seemed like a good background task when working on other stuff.

Differential Revision: D85887957