Use Int64 in Cast (#4497)
Summary:

`Cast` adds & fills any missing columns in arm data (to include all parameters in the search space) with `None`. After that, all columns are cast to the dtype matching the original parameter type. If we use `int` as the type to cast to, this errors out with `None / NaN`. The `Int64` type is more general and supports `NaN`.

Reviewed By: sdaulton

Differential Revision: D86204950