omit hypervolume trace plot when no objective thresholds (i.e., reference point) is provided (#4521)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4521

Hypervolume is a measure of multi-objective optimization progress, which corresponds to the volume enclosed by the Pareto Frontier and a reference point, set by the user via ObjectiveThresholds in Ax.

Currently, if there is no provided reference point, Ax infers the reference point based on the observed data in order to perform optimization, however its position changes as the experiment progresses and more data is observed.

This plot currently produces all zeros when no reference point is provided, which is misleading. For this reason, this diff omits this plot in this case. We should eventually decide what we want to show the user in this situation (T244212440).

Reviewed By: lena-kashtelyan

Differential Revision: D84190250

fbshipit-source-id: 2ff4a10392a814aa116e17e9301f0022aeacb1a7