Add n_best_trials_to_complete to PercentileEarlyStoppingStrategy (#4512)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4512

This diff adds a new parameter `n_best_trials_to_complete` to `PercentileEarlyStoppingStrategy` that allows protecting the top-`n_best_trials_to_complete` performing trials from early stopping, even if they fall below the percentile threshold.

The motivation for this change is to ensure that the best performing trials are allowed to run to completion, preventing the premature termination of promising trials that might temporarily perform below the percentile but still represent the best candidates in the experiment.

Reviewed By: Balandat

Differential Revision: D86345703

fbshipit-source-id: c7de2908da0a59924495038801a1e148eea3e8f0