Make `lookup_data(trial_indices=[])` return empty data (#4534)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4534

Currently, if you pass trial_indices=[], you get data for all trials because there is a check for if trial_indices rather than if trial_indices is None. This seems bad -- imagine doing `experiment.lookup_data({trial.index for trial in experiment.completed_trials})` on an experiment with no completed trials. You would get data for all the trials and they would not be completed.

Reviewed By: Balandat

Differential Revision: D86977491

fbshipit-source-id: 98f7a0cb464ec37d29528fb152930ecaca2d4465