`AddFeasibility` transform
Summary:
Transform that adds failure-awareness capability to Ax optimization.

This transform enables Ax to learn from deterministic trial failures (ABANDONED trials) and avoid sampling similar parameter configurations that are likely to fail. It achieves this by:

1. Adding a "is_feasible" metric to experiment data based on trial status
       - ABANDONED trials get feasibility value of 0.0 (infeasible)
       - Other trials get feasibility value of 1.0 (feasible)

2. Adding a feasibility constraint to the optimization config
       - The constraint enforces P(is_feasible) >= threshold
       - This guides the acquisition function to avoid infeasible regions


NOTE: We should maybe pick a different word than "feasibility" to not be confused with feasibility in the sense of not violating user-specified outcome constraints.

Differential Revision: D85185246