Update get_trace to work for preference learning experiments (#4553)
Summary:

This diff adds a new property to the `Experiment` class called `is_preference_learning_problem`. This property checks if the experiment is a preference learning (BOPE) experiment by checking if the optimization config is a `PreferenceOptimizationConfig` or if there is a `PE_EXPERIMENT` (preference exploration) auxiliary experiment attached. This property is useful for identifying preference learning experiments in Ax.

Differential Revision: D87347126