Introduce safeguards to gate early-stopping strategies (#4496)
Summary:

This diff introduces the necessary changes to support implementing safeguard logic that gates early-stopping decisions based on diagnostic checks. The safety check is defined as an abstract method in `BaseEarlyStoppingStrategy` which each subclass will be responsible for implementing. Subsequent diffs in this stack will implement concrete safety checks for subclasses, beginning with `PercentileEarlyStoppingStrategy`

Reviewed By: saitcakmak

Differential Revision: D86108540