Overwrite specified model with HeterogeneousMTGP when needed (#4561)
Summary:

If there are heterogeneous search spaces in the multi-task datasets, using a different model class will lead to errors. This makes it so that we overwrite the model class (with a warning) rather than leavivng it to error out.

Also makes sure that `Normalize` transform with `bounds=None` is used whenever `HeterogeneousMTGP` is selected.

Differential Revision: D87457028