Overwrite specified model with HeterogeneousMTGP when needed (#4561)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4561

If there are heterogeneous search spaces in the multi-task datasets, using a different model class will lead to errors. This makes it so that we overwrite the model class (with a warning) rather than leavivng it to error out.

Also makes sure that `Normalize` transform with `bounds=None` is used whenever `HeterogeneousMTGP` is selected.

Reviewed By: sdaulton

Differential Revision: D87457028

fbshipit-source-id: f646017e0ebe8a3fc5a7adb1a6222c82d40b6bc5