Add `is_bope_problem` property to `Experiment` (#4552)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4552

This diff adds a new property to the `Experiment` class called `is_bope_problem`. This property checks if the experiment is a preference learning (BOPE) experiment by checking if the optimization config is a `PreferenceOptimizationConfig` or if there is a `PE_EXPERIMENT` (preference exploration) auxiliary experiment attached. This property is useful for identifying BOPE (Bayesian Optimization Preference Exploration) experiments in Ax.

Reviewed By: ItsMrLin

Differential Revision: D87341771

fbshipit-source-id: c517fb034ecc61fc30578c8b93db6d291b8962d6