Create IndividualConstraintsFeasibilityAnalysis (#4522)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4522

Create `IndividualConstraintsFeasibilityAnalysis`. This is similar to `ConstraintsFeasibilityAnalysis`, except this look at constraints on an individual-basis and can provide more actionable feedback on which specific constraints are overly restrictive.

Currently, the heuristic used in this Analysis to deem an arm overly restrictive is pretty simple:

If `RESTRICTIVE_ARM_FRACTION` of arms are below `RESTRICTIVE_P_FEAS_THRESHOLD` for a given constraint, then that constraint is classified as overly-restrictive. Both `RESTRICTIVE_ARM_FRACTION` and `RESTRICTIVE_P_FEAS_THRESHOLD` are 0.5 as of writing this summary

Reviewed By: mgarrard

Differential Revision: D86263537

fbshipit-source-id: ed302f4cc5ab0b436c3b2e2c421f5ed46aff571a