Status-quo sem should be 0 for relative arm effects plots (#4585)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4585

Previously, the uncertainty from the sq arm was incorrectly propagated to the control arm, as though it were any other arm. Since the sq arm is the control arm, their difference should be 0 with exactly 0 uncertainty. This adds no information, however, so this diff adds functionality to exclude this arm.

## Implementation
- Updated `_relativize_df_with_sq` to accept an optional `status_quo_name` parameter
- After relativization, explicitly set the status quo arm's mean to 0.0 and sem to 0.0
- Updated `relativize_data` to retrieve and pass the status quo name
- Updated tests to verify the status quo has sem=0.0 after relativization
- Updated arm effects plotting code to omit sq point from xy scatterplot data.

Reviewed By: Balandat

Differential Revision: D87553777

fbshipit-source-id: 32f0094d9207215d239d3f8d69e3a6ffd36ffca4