Update CenterGenerationNode to transition if center has already been generated (#4583)
Summary:

The default `transition_criteria` for `CenterGenerationNode` is set to `AutoTransitionAfterGen`, which will transition to the next `GenerationNode` after generating a candidate. This works well in the standard setting where `CenterGenerationNode` generates the first candidate of an experiment.

However, if a user adds custom existing trials to the experiment, which already contain the center of the search space, the current setup will fail. In particular, `CenterGenerationNode` will not be able to generate a new candidate, because it already exists, such that its search space is exhausted. However, it will also not be able to transition to the next `GenerationNode` because it did not generate any candidate. This makes it silently call a fallback Sobol generator, which is not intended.

Instead, a better approach would be for `CenterGenerationNode` to transition to the next `GenerationNode` if its search space is exhausted, i.e. the center of the search space is already present in the existing set of trials or if it is an infeasible point.

Reviewed By: esantorella

Differential Revision: D87797277