GraphvizAnalysisCard and HierarchicalSearchSpaceGraph (#4616)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4616

Introduces GraphvizAnalysisCard which holds a DOT string in its blob attr, has a method for retrieving its graphviz.Digraph, and can render its graph as html via a svg representation. This will be generally useful for a number of visualizations, including a visualizer for the GS cc mgarrard

Also implements HierarchicalSearchSpaceGraph which will be useful now that we allow more complex trees in HSS.

This came together really quick just nuder an hour, more of a proof of  concept than anything.

Reviewed By: mgarrard

Differential Revision: D87485957

fbshipit-source-id: 83b00e680c94671d9f95aedfe7427316dfa4da1f