Remove `trial_indices_to_ignore` argument from ESS (#4626)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4626

This argument is pretty much redundant with `min_curves`. If we know that we want to not stop first k trials, we can set `min_curves=k`. Beyond that, we don't really know what the trials are going to be, so I don't see a use case where we could specify them to be not-stopped.

Storage code is also updated to discard the kwarg from json objects.

Reviewed By: ltiao

Differential Revision: D88502023

fbshipit-source-id: aeeafcd87c292182a1d4c468e792f91ff49a846d