Add expect_relativized_outcomes flag to PreferenceOptimizationConfig (#4632)
Summary:

Adds an `expect_relativized_outcomes` boolean flag to `PreferenceOptimizationConfig` to validate that outcomes are in the expected scale (relative percentage vs. absolute) before reaching the preference model. This flag will be used in later steps by the adapter layer to validate that appropriate transform pipelines are in place.

meta: although we expect relativized outcomes in PTS where BOPE is primarily used, it is a natural default to expect the outcomes to be on their raw, original scale. This is part of [the plan](https://docs.google.com/document/d/185KQovo9WAkBsIzpERZR6apRg6lbayUI0o1wZjF9EnY/edit?usp=sharing) to replace PLBO node with regular MBM node + PrefOptConfig in PTS

Reviewed By: Balandat

Differential Revision: D87721201