Fix broken OSS tests for python 3.12 from new scikit 1.8.0 release
Summary:
The recent release of scikit-learn to 1.8.0 was picked up in Ax which is causing some tests to break. 

The only difference in package versions between the newly failing tests and the old passing ones from the nightly run is the scikit-learn library. 
Pinning the scikit-learn to the most recent package version to resolve.

Differential Revision: D88866173