Rename model_state_after_gen -> generator_state_after_gen (#4666)
Summary:

Renames `model_state_after_gen` -> `generation_state_after_gen` across the code base. This includes a couple helper methods used to extract it, and the attribute and related storage for `GeneratorRun`. For SQA storage, the column name is not modified (upcoming storage refactor should clear it up) but the encoder & decoder are updated to serialize / deserialize it correctly.

Differential Revision: D89292754