Rename model_state_after_gen -> generator_state_after_gen (#4666)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4666

Renames `model_state_after_gen` -> `generation_state_after_gen` across the code base. This includes a couple helper methods used to extract it, and the attribute and related storage for `GeneratorRun`. For SQA storage, the column name is not modified (upcoming storage refactor should clear it up) but the encoder & decoder are updated to serialize / deserialize it correctly.

Reviewed By: Balandat

Differential Revision: D89292754

fbshipit-source-id: 6a07cc8d01bb5eaf3c5c3e7106027a27367bd152