Fix dependent handling in Log transform
Summary:
`Log` transform was ignoring `dependents` attribute of the `ChoiceParameters` being transformed. This hasn't been an issue so far since we did not support >2 values in HSS GS, but we should fix it for future usage.

This diff transforms the dependents in the same way it transforms the parameter values.

Differential Revision:
D89405948

Privacy Context Container: L1307644