Fix dependent handling in Log transform (#4679)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4679

`Log` transform was ignoring `dependents` attribute of the `ChoiceParameters` being transformed. This hasn't been an issue so far since we did not support >2 values in HSS GS, but we should fix it for future usage.

This diff transforms the dependents in the same way it transforms the parameter values.

Reviewed By: Balandat

Differential Revision:
D89405948

Privacy Context Container: L1307644

fbshipit-source-id: 4755e49e5d81c4c6ddf353c07dc217745365cea3