Make all fetch timestamps 0 (#4520)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4520

Makes all timestamps in `experiment._data_by_trial` 0 by updating
* json decoder
* SQA decoder
* Logic when attaching data

Important note: It is critical that with this change, it is impossible for `_data_by_trial` to be in the old format (with nonzero timestamps) once data is in memory; it can only be in the old format when saved as JSON or sqlite. Have we hit everywhere that data can be created? I removed a unit test that manually forced data to have the legacy behavior, because attempting to save data in that format will not work.

I added a check for invalidly formatted data in `lookup_data_for_trial` (which winds up called by many functions) so that there will be an informative error if data does wind up in an invalid format, but this really shouldn't happen.

Reviewed By: saitcakmak

Differential Revision: D86521191

fbshipit-source-id: 8502f53a62175da339eae52f16857ba9d3ca5dff