Recompute DerivedParameter value in Cast (#4685)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4685

This recomputes `DerivedParameter` values in `Cast.untransform_observation_features`.

This handles the case where there is a `DerivedParameter` that depends on a `RangeParameter` that has `digits` set. Without this diff, when untransforming `ObservationFeatures`, the `DerivedParameter` value would be computed using the `RangeParameter` value before it is rounded. This means that the untransfromed `ObservationFeatures` would not be in the search space. This diff fixes the issue by re-computing the `DerivedParameter` value in `Cast` after casting the other parameters.

Moving `unwrap_observation_data` to its was own file is to avoid a circular import in the original implementation, but it seems nice to keep that anyway.

Reviewed By: saitcakmak

Differential Revision: D89500685

fbshipit-source-id: 90898d1172232e01710793b17a36c8557e26ed75