Remove Metric.data_constructor (#4713)
Summary:

`Metric.data_constructor` is no longer used following D89689313.

Changes:
* Remove `Metric.data_constructor` class attribute
* Change `Metric.has_map_data` from a property into a class attribute based on whether the metric is a MapMetric

The longer-term plan is to allow (and eventually require) all metrics to return `MapData`); instead of checking whether the metric has map data, we should add an attribute `step_should_be_modeled` that tells whether that data is relevant for modeling.  This will allow for merging `Metric` and `MapMetric` as well as `Data` and `MapData`

Reviewed By: lena-kashtelyan

Differential Revision: D89748011